Revert "Working on drone controlling script."

+++ Drone_Server.py
@@ -59,8 +59,8 @@ trace_loop = False
 all_stop = False
 less_turn_percent = 0.35
 right_motor_less_pwm_percent = 0.35
-max_left_pwm = (1500 / right_motor_less_pwm_percent) / less_turn_percent
-max_right_pwm = max_left_pwm * (1 - right_motor_pwm)
+max_left_pwm = int((1500 / right_motor_less_pwm_percent) / less_turn_percent)
+max_right_pwm = int(max_left_pwm * (1 - right_motor_pwm))
 max_left_turn_pwm = 1500
 max_right_turn_pwm = 1500
